Changed Excel files and updated it on the index.txt
Changed the outputCrimes to print out the tasks:
i) were successful – i.e. resulted in arrest, caution or penalty AND the
outcome related to the stop.
ii) were partially successful – i.e. resulted in arrest, caution or
penalty, BUT the outcome DID NOT
relate to the stop
iii) were unsuccessful – i.e. a stop occurred with no further action